Finalized Portfolio Website
I went through and updated the tables in the portfolio section and I also correctly directed links to all teh appropriate places and on top of this, I have also updated the photos in each section to be appropriate for the section it is in.

I also added links and correctly directed buttons in the page to the correct social accoutns and added a couple to the account to assist in others being able to get ahold of me easier.

I added a "Im a human checkbox as well which later I will update with javascript to ensure the form is only submitted once that is checked.